+++ app/models/search/search_subject.rb
@@ -19,8 +19,8 @@ class SearchSubject
   def unify_token_operator
     @token.operator = \
       case @token.operator
-      when '!=', '!'
-        '!'
+      when "!=", "!"
+        "!"
       else
         @token.operator
       end
@@ -28,17 +28,17 @@ class SearchSubject
 
   def subjects
     case @token.key
-    when 'adverse-events'
+    when "adverse-events"
       filter_adverse_events
-    when 'comments'
+    when "comments"
       filter_comments
-    when 'designs'
+    when "designs"
       filter_designs
-    when 'events'
+    when "events"
       filter_events
-    when 'files'
+    when "files"
       filter_files
-    when 'randomized'
+    when "randomized"
       filter_randomized
     else
       event = find_event
@@ -47,10 +47,10 @@ class SearchSubject
       if event && design
         subjects = @scope.joins(:subject_events).where(subject_events: { event: event })
         case modifier
-        when 'missing'
+        when "missing"
           subjects.joins(subject_events: :sheets)
                   .where(subject_events: { event: event, sheets: { design: design, missing: true } })
-        when 'unentered'
+        when "unentered"
           subjects.where.not(
             id: subjects.joins(subject_events: :sheets)
                         .where(subject_events: { event: event, sheets: { design: design, missing: [true, false] } })
@@ -67,9 +67,11 @@ class SearchSubject
   end
 
   def filter_adverse_events
-    if @token.value == 'open'
+    if !@project.unblinded?(current_user)
+      @scope.none
+    elsif @token.value == "open"
       @scope.open_aes
-    elsif @token.value == 'closed'
+    elsif @token.value == "closed"
       @scope.closed_aes
     elsif %w(missing !).include?(@token.operator)
       @scope.where.not(id: @scope.any_aes.select(:id))
@@ -119,7 +121,7 @@ class SearchSubject
   end
 
   def filter_randomized
-    if @token.operator == '!'
+    if @token.operator == "!"
       @scope.unrandomized
     else
       @scope.randomized
@@ -132,7 +134,7 @@ class SearchSubject
         @project.designs
       else
         @project.designs.where(
-          'slug ilike any (array[?]) or id IN (?)',
+          "slug ilike any (array[?]) or id IN (?)",
           @token.values,
           @token.values.collect(&:to_i)
         )
@@ -145,7 +147,7 @@ class SearchSubject
         @project.events
       else
         @project.events.where(
-          'slug ilike any (array[?]) or id IN (?)',
+          "slug ilike any (array[?]) or id IN (?)",
           @token.values,
           @token.values.collect(&:to_i)
         )
@@ -153,14 +155,14 @@ class SearchSubject
   end
 
   def find_event
-    @project.events.find_by('slug ilike ? or id = ?', @token.key, @token.key.to_i)
+    @project.events.find_by("slug ilike ? or id = ?", @token.key, @token.key.to_i)
   end
 
   def find_design_and_modifier
-    (design_slug, modifier) = @token.value.split(':', 2)
+    (design_slug, modifier) = @token.value.split(":", 2)
     design = \
       if design_slug
-        @project.designs.find_by('slug ilike ? or id = ?', design_slug, design_slug.to_i)
+        @project.designs.find_by("slug ilike ? or id = ?", design_slug, design_slug.to_i)
       end
     [design, modifier]
   end