+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ### Bug Fix
 - Variables that exist on multiple designs no longer produce multiple columns in
   CSV exports
+- Blinded project members can no longer filter subjects by adverse events
 
 ## 0.53.0 (May 25, 2017)